[REFACTOR] Move AnyToString to common
